requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
munkres==1.1.4
numpy==2.4.0
packaging==25.0
pillow==11.3.0
pip==25.3
pyparsing==3.3.1
PySide6==6.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
setuptools==80.9.0
shiboken6==6.7.2
six==1.17.0
tornado==6.5.4
tqdm==4.67.1
unicodedata2==17.0.0
websocket-client==1.9.0
wheel==0.45.1
